+++ lib/arjdbc/db2/as400.rb
@@ -64,8 +64,11 @@ module ArJdbc
         @connection.execute_update "call qsys.qcmdexc('QSYS/CHGJOB INQMSGRPY(*SYSRPYL)',0000000031.00000)"
         @connection.execute_update "call qsys.qcmdexc('ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY(''I'')',0000000045.00000)"
       rescue Exception => e
-        raise "Could not call CHGJOB INQMSGRPY(*SYSRPYL) and ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY('I').\n" +
-              "Do you have authority to do this?\n\n#{e.inspect}"
+        # Sometimes the exception can be a false-positive when we already ran the second command
+        unless e.message.include? 'Sequence number 9876 already defined in system reply list.'
+          raise "Could not call CHGJOB INQMSGRPY(*SYSRPYL) and ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY('I').\n" +
+                    "Do you have authority to do this?\n\n#{e.inspect}"
+        end
       end
 
       begin
@@ -102,6 +105,11 @@ module ArJdbc
     # @deprecated no longer used
     def as400?
       true
+    end
+
+    # AS400 does not support TRUNCATE command. Deleting everything from the table is should be close enough
+    def truncate(table_name, name = nil)
+      @connection.execute_update "DELETE FROM #{table_name}"
     end
 
     private