+++ lib/arjdbc/jdbc/adapter.rb
@@ -264,13 +264,16 @@ module ActiveRecord
 
       # @override
       def reconnect!
+        super # clear_cache! && reset_transaction
         @connection.reconnect! # handles adapter.configure_connection
         @connection
       end
 
       # @override
       def disconnect!
-        @connection.disconnect! if @connection
+        super # clear_cache! && reset_transaction
+        return unless @connection
+        @connection.disconnect!
       end
 
       # @override